CB-5546 Add reverse DNS zone test, refactor child env test
