+++ integration-test/src/main/java/com/sequenceiq/it/cloudbreak/testcase/mock/EnvironmentTest.java
@@ -3,7 +3,9 @@ package com.sequenceiq.it.cloudbreak.testcase.mock;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.HashSet;
+import java.util.List;
 import java.util.Set;
+import java.util.stream.Collectors;
 
 import javax.inject.Inject;
 import javax.ws.rs.BadRequestException;
@@ -12,7 +14,6 @@ import javax.ws.rs.NotFoundException;
 import org.apache.commons.lang3.StringUtils;
 import org.testng.annotations.Test;
 
-import com.sequenceiq.environment.api.v1.environment.model.response.EnvironmentStatus;
 import com.sequenceiq.environment.api.v1.environment.model.response.SimpleEnvironmentResponse;
 import com.sequenceiq.it.cloudbreak.EnvironmentClient;
 import com.sequenceiq.it.cloudbreak.client.CredentialTestClient;
@@ -26,15 +27,9 @@ import com.sequenceiq.it.cloudbreak.dto.environment.EnvironmentTestDto;
 import com.sequenceiq.it.cloudbreak.exception.TestFailException;
 import com.sequenceiq.it.cloudbreak.testcase.AbstractIntegrationTest;
 
-import static java.util.Objects.isNull;
-
 public class EnvironmentTest extends AbstractIntegrationTest {
 
     private static final Set<String> INVALID_REGION = new HashSet<>(Collections.singletonList("MockRegion"));
-
-    private static final String PARENT_ENVIRONMENT = "parent";
-
-    private static final String CHILD_ENVIRONMENT = "child";
 
     @Inject
     private EnvironmentTestClient environmentTestClient;
@@ -60,7 +55,7 @@ public class EnvironmentTest extends AbstractIntegrationTest {
                 .given(EnvironmentTestDto.class)
                 .when(environmentTestClient.create())
                 .when(environmentTestClient.list())
-                .then(this::checkEnvIsListedByNameAndParentName)
+                .then(this::checkEnvIsListed)
                 .validate();
     }
 
@@ -126,7 +121,7 @@ public class EnvironmentTest extends AbstractIntegrationTest {
                 .when(environmentTestClient.create())
                 .when(environmentTestClient.describe())
                 .when(environmentTestClient.list())
-                .then(this::checkEnvIsListedByNameAndParentName)
+                .then(this::checkEnvIsListed)
                 .when(environmentTestClient.delete())
                 .validate();
     }
@@ -162,153 +157,15 @@ public class EnvironmentTest extends AbstractIntegrationTest {
                 .validate();
     }
 
-    @Test(dataProvider = TEST_CONTEXT_WITH_MOCK)
-    @Description(
-            given = "there is a running cloudbreak",
-            when = "valid create environment requests are sent to create parent environment first and then a child one",
-            then = "environments should be created and parent environment should be referenced in the child environment")
-    public void testCreateParentChildEnvironment(TestContext testContext) {
-        testContext
-                .given(CredentialTestDto.class)
-                .when(credentialTestClient.create())
-                .given(PARENT_ENVIRONMENT, EnvironmentTestDto.class)
-                .when(environmentTestClient.create())
-                .when(environmentTestClient.list())
-                .then(this::checkEnvIsListedByNameAndParentName)
-                .await(EnvironmentStatus.AVAILABLE)
-                .given(CHILD_ENVIRONMENT, EnvironmentTestDto.class)
-                .withParentEnvironment(RunningParameter.key(PARENT_ENVIRONMENT))
-                .when(environmentTestClient.create())
-                .when(environmentTestClient.list())
-                .then(this::checkEnvIsListedByNameAndParentName)
-                .validate();
-    }
-
-    @Test(dataProvider = TEST_CONTEXT_WITH_MOCK)
-    @Description(
-            given = "there is an available child environment with a referenced available parent environment",
-            when = "child create request is sent but parent environment is a child environment",
-            then = "a BadRequestException should be returned")
-    public void testCreateParentChildEnvironmentWhereParentIsAChild(TestContext testContext) {
-        String forbiddenKey = resourcePropertyProvider().getName();
-        testContext
-                .given(CredentialTestDto.class)
-                .when(credentialTestClient.create())
-                .given(PARENT_ENVIRONMENT, EnvironmentTestDto.class)
-                .when(environmentTestClient.create())
-                .await(EnvironmentStatus.AVAILABLE)
-                .given(CHILD_ENVIRONMENT, EnvironmentTestDto.class)
-                .withParentEnvironment(RunningParameter.key(PARENT_ENVIRONMENT))
-                .when(environmentTestClient.create())
-                .await(EnvironmentStatus.AVAILABLE)
-                .given(EnvironmentTestDto.class)
-                .withParentEnvironment(RunningParameter.key(CHILD_ENVIRONMENT))
-                .when(environmentTestClient.create(), RunningParameter.key(forbiddenKey))
-                .expect(BadRequestException.class, RunningParameter.key(forbiddenKey))
-                .validate();
-    }
-
-    @Test(dataProvider = TEST_CONTEXT_WITH_MOCK)
-    @Description(
-            given = "there is a running cloudbreak",
-            when = "child create request is sent but parent is not created yet",
-            then = "a BadRequestException should be returned")
-    public void testCreateChildWithoutParentEnvironment(TestContext testContext) {
-        String forbiddenKey = resourcePropertyProvider().getName();
-        testContext
-                .given(CredentialTestDto.class)
-                .when(credentialTestClient.create())
-                .given(PARENT_ENVIRONMENT, EnvironmentTestDto.class)
-                .given(CHILD_ENVIRONMENT, EnvironmentTestDto.class)
-                .withParentEnvironment(RunningParameter.key(PARENT_ENVIRONMENT))
-                .when(environmentTestClient.create(), RunningParameter.key(forbiddenKey))
-                .expect(BadRequestException.class, RunningParameter.key(forbiddenKey))
-                .validate();
-    }
-
-    @Test(dataProvider = TEST_CONTEXT_WITH_MOCK)
-    @Description(
-            given = "there is an available child environment with a referenced available parent environment",
-            when = "a delete request is sent for the parent environment",
-            then = "a BadRequestException should be returned")
-    public void testDeleteParentEnvironmentWithExistingChild(TestContext testContext) {
-        String forbiddenKey = resourcePropertyProvider().getName();
-        testContext
-                .given(CredentialTestDto.class)
-                .when(credentialTestClient.create())
-                .given(PARENT_ENVIRONMENT, EnvironmentTestDto.class)
-                .when(environmentTestClient.create())
-                .await(EnvironmentStatus.AVAILABLE)
-                .given(CHILD_ENVIRONMENT, EnvironmentTestDto.class)
-                .withParentEnvironment(RunningParameter.key(PARENT_ENVIRONMENT))
-                .when(environmentTestClient.create())
-                .await(EnvironmentStatus.AVAILABLE)
-                .given(PARENT_ENVIRONMENT, EnvironmentTestDto.class)
-                .when(environmentTestClient.deleteByName(), RunningParameter.key(forbiddenKey))
-                .expect(BadRequestException.class, RunningParameter.key(forbiddenKey))
-                .validate();
-    }
-
-    @Test(dataProvider = TEST_CONTEXT_WITH_MOCK)
-    @Description(
-            given = "there is an available child environment with a referenced available parent environment",
-            when = "a delete request is sent for the child environment",
-            then = "the child environment should be deleted")
-    public void testDeleteChildEnvironment(TestContext testContext) {
-        testContext
-                .given(CredentialTestDto.class)
-                .when(credentialTestClient.create())
-                .given(PARENT_ENVIRONMENT, EnvironmentTestDto.class)
-                .when(environmentTestClient.create())
-                .await(EnvironmentStatus.AVAILABLE)
-                .given(CHILD_ENVIRONMENT, EnvironmentTestDto.class)
-                .withParentEnvironment(RunningParameter.key(PARENT_ENVIRONMENT))
-                .when(environmentTestClient.create())
-                .await(EnvironmentStatus.AVAILABLE)
-                .when(environmentTestClient.deleteByName())
-                .await(EnvironmentStatus.ARCHIVED)
-                .when(environmentTestClient.list())
-                .then(this::checkEnvIsNotListedByNameAndParentName)
-                .validate();
-    }
-
-    private EnvironmentTestDto checkEnvIsListedByNameAndParentName(TestContext testContext,
-            EnvironmentTestDto environment,
-            EnvironmentClient environmentClient) {
+    private EnvironmentTestDto checkEnvIsListed(TestContext testContext, EnvironmentTestDto environment, EnvironmentClient environmentClient) {
         Collection<SimpleEnvironmentResponse> simpleEnvironmentV4Respons = environment.getResponseSimpleEnvSet();
-        if (isNull(simpleEnvironmentV4Respons)) {
-            throw new TestFailException("Environment list response is missing.");
-        }
-        boolean listed = simpleEnvironmentV4Respons.stream()
-                .anyMatch(environmentResponse -> nameEquals(environment, environmentResponse) && parentNameEquals(environment, environmentResponse));
-        if (!listed) {
+        List<SimpleEnvironmentResponse> result = simpleEnvironmentV4Respons.stream()
+                .filter(env -> environment.getName().equals(env.getName()))
+                .collect(Collectors.toList());
+        if (result.isEmpty()) {
             throw new TestFailException("Environment is not listed");
         }
         return environment;
-    }
-
-    private EnvironmentTestDto checkEnvIsNotListedByNameAndParentName(TestContext testContext,
-            EnvironmentTestDto environment,
-            EnvironmentClient environmentClient) {
-        Collection<SimpleEnvironmentResponse> simpleEnvironmentV4Respons = environment.getResponseSimpleEnvSet();
-        if (isNull(simpleEnvironmentV4Respons)) {
-            throw new TestFailException("Environment list response is missing.");
-        }
-        boolean listed = simpleEnvironmentV4Respons.stream()
-                .anyMatch(environmentResponse -> nameEquals(environment, environmentResponse) && parentNameEquals(environment, environmentResponse));
-        if (listed) {
-            throw new TestFailException("Environment is listed");
-        }
-        return environment;
-    }
-
-    private boolean nameEquals(EnvironmentTestDto environment, SimpleEnvironmentResponse environmentResponse) {
-        return environment.getName().equals(environmentResponse.getName());
-    }
-
-    private boolean parentNameEquals(EnvironmentTestDto environment, SimpleEnvironmentResponse environmentResponse) {
-        return isNull(environment.getParentEnvironmentName()) && isNull(environmentResponse.getParentEnvironmentName()) ||
-                environment.getParentEnvironmentName().equals(environmentResponse.getParentEnvironmentName());
     }
 
     private static EnvironmentTestDto checkEnvironmentCrnIsNotEmpty(TestContext testContext,
